[CLEAN] package objectからtraitに切り出す
